requirements.txt
requests==2.32.3
numpy==1.26.4
fake-useragent==1.5.1
python-dotenv==1.0.1
mistralai==1.1.0
openai-whisper==20240930
vosk==0.3.45
SpeechRecognition==3.11.0
transformers==4.46.1
sentencepiece==0.2.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1